+++ src/appearance/theme.rs
@@ -106,6 +106,13 @@ impl iced::theme::Base for Theme {
 
 impl combo_box::Catalog for Theme {}
 
+pub fn line_height(config: &config::Font) -> LineHeight {
+    config
+        .line_height
+        .map(LineHeight::Relative)
+        .unwrap_or_default()
+}
+
 pub fn resolve_line_height(config: &config::Font) -> f32 {
     LineHeight::default()
         .to_absolute(