+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Added:
 - Setting to style unread query buffers as highlights in the sidebar (`sidebar.unread_indicator.query_as_highlight`)
 - Settings to configure card and image preview dimensions 
 - Support for transparent background.
+- Setting to configure font line height (`font.line_height`)
 
 Fixed: